wip: add order modeling

+++ src/main/java/com/commerceplatform/api/orders/controllers/OrderController.java
@@ -1,12 +1,10 @@
 package com.commerceplatform.api.orders.controllers;
 
-import com.commerceplatform.api.orders.integrations.api.products.dtos.ProductDto;
 import com.commerceplatform.api.orders.models.jpa.OrderModel;
 import com.commerceplatform.api.orders.services.OrderService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -26,7 +24,7 @@ public class OrderController {
     @PostMapping
     public ResponseEntity<OrderModel> createOrder() {
 
-        List<Long> dto = new ArrayList<>();
+        List<Integer> dto = new ArrayList<>();
         Long customerId = 3L;
         return ResponseEntity.status(HttpStatus.OK).body(orderService.createOrder(customerId, dto));
     }
